refactor: common utils
