+++ tpch/modin_queries/queries.py
@@ -5,36 +5,19 @@ import argparse
 import traceback
 from typing import Dict
 
-import xorbits
-import xorbits.pandas as pd
-pd.set_option("show_progress", False)
+import modin
+import ray
+import modin.pandas as pd
 
-TIMINGS_FILE = "time.csv"
+from common_utils import log_time_fn, parse_common_arguments, print_result_fn
+
 dataset_dict = {}
 
 
-def append_row(solution: str, q: str, secs: float, version: str, success=True):
-    with open(TIMINGS_FILE, "a") as f:
-        if f.tell() == 0:
-            f.write("solution,version,query_no,duration[s],success\n")
-        f.write(f"{solution},{version},{q},{secs},{success}\n")
-        print(f"{solution},{version},{q},{secs},{success}")
-
-
-def load_lineitem(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_lineitem(root: str, storage_options: Dict):
     if "lineitem" not in dataset_dict:
         data_path = root + "/lineitem"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         df.L_SHIPDATE = pd.to_datetime(df.L_SHIPDATE, format="%Y-%m-%d")
         df.L_RECEIPTDATE = pd.to_datetime(df.L_RECEIPTDATE, format="%Y-%m-%d")
         df.L_COMMITDATE = pd.to_datetime(df.L_COMMITDATE, format="%Y-%m-%d")
@@ -45,20 +28,10 @@ def load_lineitem(
     return result
 
 
-def load_part(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_part(root: str, storage_options: Dict):
     if "part" not in dataset_dict:
         data_path = root + "/part"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         result = df
         dataset_dict["part"] = result
     else:
@@ -66,20 +39,10 @@ def load_part(
     return result
 
 
-def load_orders(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_orders(root: str, storage_options: Dict):
     if "orders" not in dataset_dict:
         data_path = root + "/orders"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         df.O_ORDERDATE = pd.to_datetime(df.O_ORDERDATE, format="%Y-%m-%d")
         result = df
         dataset_dict["orders"] = result
@@ -88,20 +51,10 @@ def load_orders(
     return result
 
 
-def load_customer(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_customer(root: str, storage_options: Dict):
     if "customer" not in dataset_dict:
         data_path = root + "/customer"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         result = df
         dataset_dict["customer"] = result
     else:
@@ -109,20 +62,10 @@ def load_customer(
     return result
 
 
-def load_nation(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_nation(root: str, storage_options: Dict):
     if "nation" not in dataset_dict:
         data_path = root + "/nation"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         result = df
         dataset_dict["nation"] = result
     else:
@@ -130,20 +73,10 @@ def load_nation(
     return result
 
 
-def load_region(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_region(root: str, storage_options: Dict):
     if "region" not in dataset_dict:
         data_path = root + "/region"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         result = df
         dataset_dict["region"] = result
     else:
@@ -151,20 +84,10 @@ def load_region(
     return result
 
 
-def load_supplier(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_supplier(root: str, storage_options: Dict):
     if "supplier" not in dataset_dict:
         data_path = root + "/supplier"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         result = df
         dataset_dict["supplier"] = result
     else:
@@ -172,20 +95,10 @@ def load_supplier(
     return result
 
 
-def load_partsupp(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def load_partsupp(root: str, storage_options: Dict):
     if "partsupp" not in dataset_dict:
         data_path = root + "/partsupp"
-        df = pd.read_parquet(
-            data_path,
-            use_arrow_dtype=use_arrow_dtype,
-            storage_options=storage_options,
-            gpu=gpu,
-        )
+        df = pd.read_parquet(data_path, storage_options=storage_options)
         result = df
         dataset_dict["partsupp"] = result
     else:
@@ -193,12 +106,7 @@ def load_partsupp(
     return result
 
 
-def q01(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q01(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
 
     date = pd.Timestamp("1998-09-02")
@@ -240,26 +148,22 @@ def q01(
             "L_ORDERKEY": "count",
         }
     )
+    total = (
+        total.sort_values(["L_RETURNFLAG", "L_LINESTATUS"])
+            .rename(columns={
+                "L_QUANTITY": "SUM_QTY",
+                "L_EXTENDEDPRICE": "SUM_BASE_PRICE",
+                "DISC_PRICE": "SUM_DISC_PRICE",
+                "CHARGE": "SUM_CHARGE",
+                "L_DISCOUNT": "AVG_DISC",
+                "L_ORDERKEY": "COUNT_ORDER"
+            })
+    )
 
-    # skip sort, Xorbits groupby enables sort
-    # total = total.sort_values(["L_RETURNFLAG", "L_LINESTATUS"])
-    total = total.rename(columns={
-                    "L_QUANTITY": "SUM_QTY",
-                    "L_EXTENDEDPRICE": "SUM_BASE_PRICE",
-                    "DISC_PRICE": "SUM_DISC_PRICE",
-                    "CHARGE": "SUM_CHARGE",
-                    "L_DISCOUNT": "AVG_DISC",
-                    "L_ORDERKEY": "COUNT_ORDER"
-                })
     return total
 
 
-def q02(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q02(root: str, storage_options: Dict):
     part = load_part(root, storage_options)
     partsupp = load_partsupp(root, storage_options)
     supplier = load_supplier(root, storage_options)
@@ -344,7 +248,7 @@ def q02(
             "P_MFGR",
         ],
     ]
-    min_values = merged_df.groupby("P_PARTKEY", as_index=False, sort=False)["PS_SUPPLYCOST"].min()
+    min_values = merged_df.groupby("P_PARTKEY", as_index=False)["PS_SUPPLYCOST"].min()
     min_values.columns = ["P_PARTKEY_CPY", "MIN_SUPPLYCOST"]
     merged_df = merged_df.merge(
         min_values,
@@ -366,30 +270,15 @@ def q02(
         ],
     ]
     total = total.sort_values(
-        by=[
-            "S_ACCTBAL",
-            "N_NAME",
-            "S_NAME",
-            "P_PARTKEY",
-        ],
-        ascending=[
-            False,
-            True,
-            True,
-            True,
-        ],
+        by=["S_ACCTBAL", "N_NAME", "S_NAME", "P_PARTKEY"],
+        ascending=[False, True, True, True],
     )
     total = total.head(100)
 
     return total
 
 
-def q03(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q03(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
     customer = load_customer(root, storage_options)
@@ -399,9 +288,7 @@ def q03(
     lineitem_filtered = lineitem.loc[
         :, ["L_ORDERKEY", "L_EXTENDEDPRICE", "L_DISCOUNT", "L_SHIPDATE"]
     ]
-    orders_filtered = orders.loc[
-        :, ["O_ORDERKEY", "O_CUSTKEY", "O_ORDERDATE", "O_SHIPPRIORITY"]
-    ]
+    orders_filtered = orders.loc[:, ["O_ORDERKEY", "O_CUSTKEY", "O_ORDERDATE", "O_SHIPPRIORITY"]]
     customer_filtered = customer.loc[:, ["C_MKTSEGMENT", "C_CUSTKEY"]]
     lsel = lineitem_filtered.L_SHIPDATE > date
     osel = orders_filtered.O_ORDERDATE < date
@@ -413,7 +300,7 @@ def q03(
     jn2 = jn1.merge(flineitem, left_on="O_ORDERKEY", right_on="L_ORDERKEY")
     jn2["REVENUE"] = jn2.L_EXTENDEDPRICE * (1 - jn2.L_DISCOUNT)
     total = (
-        jn2.groupby(["L_ORDERKEY", "O_ORDERDATE", "O_SHIPPRIORITY"], as_index=False, sort=False)["REVENUE"]
+        jn2.groupby(["L_ORDERKEY", "O_ORDERDATE", "O_SHIPPRIORITY"], as_index=False)["REVENUE"]
         .sum()
         .sort_values(["REVENUE"], ascending=False)
     )
@@ -425,12 +312,7 @@ def q03(
     return total
 
 
-def q04(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q04(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
 
@@ -445,18 +327,14 @@ def q04(
     total = (
         jn.groupby("O_ORDERPRIORITY", as_index=False)["O_ORDERKEY"]
         .count()
+        .sort_values(["O_ORDERPRIORITY"])
         .rename(columns={"O_ORDERKEY": "ORDER_COUNT"})
     )
 
     return total
 
 
-def q05(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q05(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
     customer = load_customer(root, storage_options)
@@ -481,18 +359,13 @@ def q05(
         jn4, left_on=["S_SUPPKEY", "S_NATIONKEY"], right_on=["L_SUPPKEY", "N_NATIONKEY"]
     )
     jn5["REVENUE"] = jn5.L_EXTENDEDPRICE * (1.0 - jn5.L_DISCOUNT)
-    gb = jn5.groupby("N_NAME", as_index=False, sort=False)["REVENUE"].sum()
+    gb = jn5.groupby("N_NAME", as_index=False)["REVENUE"].sum()
     total = gb.sort_values("REVENUE", ascending=False)
 
     return total
 
 
-def q06(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q06(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
 
     date1 = pd.Timestamp("1996-01-01")
@@ -514,12 +387,7 @@ def q06(
     return result_df
 
 
-def q07(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q07(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
     customer = load_customer(root, storage_options)
@@ -604,28 +472,19 @@ def q07(
     # concat results
     total = pd.concat([total1, total2])
 
-    total = total.groupby(["SUPP_NATION", "CUST_NATION", "L_YEAR"], as_index=False).agg(
-        REVENUE=pd.NamedAgg(column="VOLUME", aggfunc="sum")
+    total = (
+        total.groupby(["SUPP_NATION", "CUST_NATION", "L_YEAR"], as_index=False)
+            .agg(REVENUE=pd.NamedAgg(column="VOLUME", aggfunc="sum"))
+            .sort_values(
+                by=["SUPP_NATION", "CUST_NATION", "L_YEAR"],
+                ascending=[True, True, True]
+            )
     )
-    # skip sort when Xorbits groupby does sort already
-    # total = total.sort_values(
-    #     by=["SUPP_NATION", "CUST_NATION", "L_YEAR"],
-    #     ascending=[
-    #         True,
-    #         True,
-    #         True,
-    #     ],
-    # )
 
     return total
 
 
-def q08(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q08(root: str, storage_options: Dict):
     part = load_part(root, storage_options)
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
@@ -640,9 +499,7 @@ def q08(
     part_filtered = part[(part["P_TYPE"] == p_type)]
     part_filtered = part_filtered.loc[:, ["P_PARTKEY"]]
     lineitem_filtered = lineitem.loc[:, ["L_PARTKEY", "L_SUPPKEY", "L_ORDERKEY"]]
-    lineitem_filtered["VOLUME"] = lineitem["L_EXTENDEDPRICE"] * (
-        1.0 - lineitem["L_DISCOUNT"]
-    )
+    lineitem_filtered["VOLUME"] = lineitem["L_EXTENDEDPRICE"] * (1.0 - lineitem["L_DISCOUNT"])
     total = part_filtered.merge(
         lineitem_filtered, left_on="P_PARTKEY", right_on="L_PARTKEY", how="inner"
     )
@@ -668,9 +525,8 @@ def q08(
     )
     total = total.loc[:, ["VOLUME", "S_NATIONKEY", "O_YEAR", "C_NATIONKEY"]]
     n1_filtered = nation.loc[:, ["N_NATIONKEY", "N_REGIONKEY"]]
-    n2_filtered = nation.loc[:, ["N_NATIONKEY", "N_NAME"]].rename(
-        columns={"N_NAME": "NATION"}
-    )
+    n2_filtered = nation.loc[:, ["N_NATIONKEY", "N_NAME"]] \
+        .rename(columns={"N_NAME": "NATION"})
     total = total.merge(
         n1_filtered, left_on="C_NATIONKEY", right_on="N_NATIONKEY", how="inner"
     )
@@ -692,19 +548,14 @@ def q08(
         numerator = df["VOLUME"].sum()
         return numerator / demonimator
 
-    total = total.groupby("O_YEAR", as_index=False).apply(udf)
+    total = total.groupby("O_YEAR").apply(udf).reset_index()
     total.columns = ["O_YEAR", "MKT_SHARE"]
-    total = total.sort_values(by=["O_YEAR",], ascending=[True,],)
+    total = total.sort_values(by=["O_YEAR"], ascending=[True])
 
     return total
 
 
-def q09(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q09(root: str, storage_options: Dict):
     part = load_part(root, storage_options)
     partsupp = load_partsupp(root, storage_options)
     lineitem = load_lineitem(root, storage_options)
@@ -725,7 +576,7 @@ def q09(
     jn5["TMP"] = jn5.L_EXTENDEDPRICE * (1 - jn5.L_DISCOUNT) - (
         (1 * jn5.PS_SUPPLYCOST) * jn5.L_QUANTITY
     )
-    jn5["O_YEAR"] = jn5.O_ORDERDATE.apply(lambda x: x.year)
+    jn5["O_YEAR"] = jn5.O_ORDERDATE.dt.year
     gb = jn5.groupby(["N_NAME", "O_YEAR"], as_index=False)["TMP"].sum()
     total = gb.sort_values(["N_NAME", "O_YEAR"], ascending=[True, False])
     total = total.rename(columns={"TMP": "SUM_PROFIT"})
@@ -733,12 +584,7 @@ def q09(
     return total
 
 
-def q10(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q10(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
     nation = load_nation(root, storage_options)
@@ -784,12 +630,7 @@ def q10(
     return total
 
 
-def q11(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q11(root: str, storage_options: Dict):
     partsupp = load_partsupp(root, storage_options)
     supplier = load_supplier(root, storage_options)
     nation = load_nation(root, storage_options)
@@ -822,12 +663,7 @@ def q11(
     return total
 
 
-def q12(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q12(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
 
@@ -854,18 +690,14 @@ def q12(
         return ((x != "1-URGENT") & (x != "2-HIGH")).sum()
 
     total = jn.groupby("L_SHIPMODE", as_index=False)["O_ORDERPRIORITY"].agg((g1, g2))
-    # skip sort when groupby does sort already
-    # total = total.sort_values("L_SHIPMODE")
-    total = total.rename(columns={"g1": "HIGH_LINE_COUNT", "g2": "LOW_LINE_COUNT"})
+    total = (
+        total.sort_values("L_SHIPMODE")
+        .rename(columns={"g1": "HIGH_LINE_COUNT", "g2": "LOW_LINE_COUNT"})
+    )
     return total
 
 
-def q13(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q13(root: str, storage_options: Dict):
     customer = load_customer(root, storage_options)
     orders = load_orders(root, storage_options)
 
@@ -880,28 +712,20 @@ def q13(
         orders_filtered, left_on="C_CUSTKEY", right_on="O_CUSTKEY", how="left"
     )
     c_o_merged = c_o_merged.loc[:, ["C_CUSTKEY", "O_ORDERKEY"]]
-    count_df = c_o_merged.groupby(["C_CUSTKEY"], as_index=False, sort=False).agg(
+    count_df = c_o_merged.groupby(["C_CUSTKEY"], as_index=False).agg(
         C_COUNT=pd.NamedAgg(column="O_ORDERKEY", aggfunc="count")
     )
-    total = count_df.groupby(["C_COUNT"], as_index=False, sort=False).size()
+    total = count_df.groupby(["C_COUNT"], as_index=False).size()
     total.columns = ["C_COUNT", "CUSTDIST"]
     total = total.sort_values(
         by=["CUSTDIST", "C_COUNT"],
-        ascending=[
-            False,
-            False,
-        ],
+        ascending=[False, False],
     )
 
     return total
 
 
-def q14(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q14(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     part = load_part(root, storage_options)
 
@@ -912,44 +736,33 @@ def q14(
     lineitem_filtered = lineitem.loc[
         :, ["L_EXTENDEDPRICE", "L_DISCOUNT", "L_SHIPDATE", "L_PARTKEY"]
     ]
-    sel = (lineitem_filtered.L_SHIPDATE >= startDate) & (
-        lineitem_filtered.L_SHIPDATE < endDate
-    )
+    sel = (lineitem_filtered.L_SHIPDATE >= startDate) \
+        & (lineitem_filtered.L_SHIPDATE < endDate)
     flineitem = lineitem_filtered[sel]
     jn = flineitem.merge(part_filtered, left_on="L_PARTKEY", right_on="P_PARTKEY")
     jn["PROMO_REVENUE"] = jn.L_EXTENDEDPRICE * (1.0 - jn.L_DISCOUNT)
     total = (
-        jn[jn.P_TYPE.str.startswith(p_type_like)].PROMO_REVENUE.sum()
-        * 100
-        / jn.PROMO_REVENUE.sum()
+        jn[jn.P_TYPE.str.startswith(p_type_like)].PROMO_REVENUE.sum() * 100 / jn.PROMO_REVENUE.sum()
     )
 
     result_df = pd.DataFrame({"PROMO_REVENUE": [total]})
     return result_df
 
 
-def q15(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q15(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     supplier = load_supplier(root, storage_options)
 
     lineitem_filtered = lineitem[
         (lineitem["L_SHIPDATE"] >= pd.Timestamp("1996-01-01"))
-        & (
-            lineitem["L_SHIPDATE"]
-            < (pd.Timestamp("1996-01-01") + pd.DateOffset(months=3))
-        )
+        & (lineitem["L_SHIPDATE"] < (pd.Timestamp("1996-01-01") + pd.DateOffset(months=3)))
     ]
     lineitem_filtered["REVENUE_PARTS"] = lineitem_filtered["L_EXTENDEDPRICE"] * (
         1.0 - lineitem_filtered["L_DISCOUNT"]
     )
     lineitem_filtered = lineitem_filtered.loc[:, ["L_SUPPKEY", "REVENUE_PARTS"]]
     revenue_table = (
-        lineitem_filtered.groupby("L_SUPPKEY", as_index=False, sort=False)
+        lineitem_filtered.groupby("L_SUPPKEY", as_index=False)
         .agg(TOTAL_REVENUE=pd.NamedAgg(column="REVENUE_PARTS", aggfunc="sum"))
         .rename(columns={"L_SUPPKEY": "SUPPLIER_NO"})
     )
@@ -966,12 +779,7 @@ def q15(
     return total
 
 
-def q16(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q16(root: str, storage_options: Dict):
     part = load_part(root, storage_options)
     partsupp = load_partsupp(root, storage_options)
     supplier = load_supplier(root, storage_options)
@@ -1007,9 +815,8 @@ def q16(
     )
     total = total[total["S_SUPPKEY"].isna()]
     total = total.loc[:, ["P_BRAND", "P_TYPE", "P_SIZE", "PS_SUPPKEY"]]
-    total = total.groupby(["P_BRAND", "P_TYPE", "P_SIZE"], as_index=False, sort=False)[
-        "PS_SUPPKEY"
-    ].nunique()
+    total = total.groupby(["P_BRAND", "P_TYPE", "P_SIZE"], as_index=False)["PS_SUPPKEY"] \
+                .nunique()
     total.columns = ["P_BRAND", "P_TYPE", "P_SIZE", "SUPPLIER_CNT"]
     total = total.sort_values(
         by=["SUPPLIER_CNT", "P_BRAND", "P_TYPE", "P_SIZE"],
@@ -1019,12 +826,7 @@ def q16(
     return total
 
 
-def q17(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q17(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     part = load_part(root, storage_options)
 
@@ -1041,9 +843,9 @@ def q17(
         :, ["L_QUANTITY", "L_EXTENDEDPRICE", "P_PARTKEY"]
     ]
     lineitem_filtered = lineitem.loc[:, ["L_PARTKEY", "L_QUANTITY"]]
-    lineitem_avg = lineitem_filtered.groupby(
-        ["L_PARTKEY"], as_index=False, sort=False
-    ).agg(avg=pd.NamedAgg(column="L_QUANTITY", aggfunc="mean"))
+    lineitem_avg = lineitem_filtered.groupby(["L_PARTKEY"], as_index=False).agg(
+        avg=pd.NamedAgg(column="L_QUANTITY", aggfunc="mean")
+    )
     lineitem_avg["avg"] = 0.2 * lineitem_avg["avg"]
     lineitem_avg = lineitem_avg.loc[:, ["L_PARTKEY", "avg"]]
     total = line_part_merge.merge(
@@ -1055,12 +857,7 @@ def q17(
     return total
 
 
-def q18(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q18(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
     customer = load_customer(root, storage_options)
@@ -1073,7 +870,6 @@ def q18(
     gb2 = jn2.groupby(
         ["C_NAME", "C_CUSTKEY", "O_ORDERKEY", "O_ORDERDATE", "O_TOTALPRICE"],
         as_index=False,
-        sort=False,
     )["L_QUANTITY"].sum()
     total = gb2.sort_values(["O_TOTALPRICE", "O_ORDERDATE"], ascending=[False, True])
     total = total.head(100)
@@ -1081,12 +877,7 @@ def q18(
     return total
 
 
-def q19(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q19(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     part = load_part(root, storage_options)
 
@@ -1183,18 +974,12 @@ def q19(
     )
     jn = jn[jnsel]
     result_value = (jn.L_EXTENDEDPRICE * (1.0 - jn.L_DISCOUNT)).sum()
-
     result_df = pd.DataFrame({"REVENUE": [result_value]})
 
     return result_df
 
 
-def q20(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q20(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     part = load_part(root, storage_options)
     nation = load_nation(root, storage_options)
@@ -1216,9 +1001,9 @@ def q20(
         left_on=["PS_PARTKEY", "PS_SUPPKEY"],
         right_on=["L_PARTKEY", "L_SUPPKEY"],
     )
-    gb = jn2.groupby(
-        ["PS_PARTKEY", "PS_SUPPKEY", "PS_AVAILQTY"], as_index=False, sort=False
-    )["L_QUANTITY"].sum()
+    gb = jn2.groupby(["PS_PARTKEY", "PS_SUPPKEY", "PS_AVAILQTY"], as_index=False)[
+        "L_QUANTITY"
+    ].sum()
     gbsel = gb.PS_AVAILQTY > (0.5 * gb.L_QUANTITY)
     fgb = gb[gbsel]
     jn3 = fgb.merge(supplier, left_on="PS_SUPPKEY", right_on="S_SUPPKEY")
@@ -1229,12 +1014,7 @@ def q20(
     return total
 
 
-def q21(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q21(root: str, storage_options: Dict):
     lineitem = load_lineitem(root, storage_options)
     orders = load_orders(root, storage_options)
     supplier = load_supplier(root, storage_options)
@@ -1248,7 +1028,7 @@ def q21(
     # Exists
     lineitem_orderkeys = (
         lineitem_filtered.loc[:, ["L_ORDERKEY", "L_SUPPKEY"]]
-        .groupby("L_ORDERKEY", as_index=False, sort=False)["L_SUPPKEY"]
+        .groupby("L_ORDERKEY", as_index=False)["L_SUPPKEY"]
         .nunique()
     )
     lineitem_orderkeys.columns = ["L_ORDERKEY", "nunique_col"]
@@ -1267,9 +1047,10 @@ def q21(
     )
 
     # Not Exists: Check the exists condition isn't still satisfied on the output.
-    lineitem_orderkeys = lineitem_filtered.groupby(
-        "L_ORDERKEY", as_index=False, sort=False
-    )["L_SUPPKEY"].nunique()
+    lineitem_orderkeys = (
+        lineitem_filtered.groupby("L_ORDERKEY", as_index=False)["L_SUPPKEY"]
+        .nunique()
+    )
     lineitem_orderkeys.columns = ["L_ORDERKEY", "nunique_col"]
     lineitem_orderkeys = lineitem_orderkeys[lineitem_orderkeys["nunique_col"] == 1]
     lineitem_orderkeys = lineitem_orderkeys.loc[:, ["L_ORDERKEY"]]
@@ -1298,7 +1079,7 @@ def q21(
         nation_filtered, left_on="S_NATIONKEY", right_on="N_NATIONKEY", how="inner"
     )
     total = total.loc[:, ["S_NAME"]]
-    total = total.groupby("S_NAME", as_index=False, sort=False).size()
+    total = total.groupby("S_NAME", as_index=False).size()
     total.columns = ["S_NAME", "NUMWAIT"]
     total = total.sort_values(by=["NUMWAIT", "S_NAME"], ascending=[False, True])
     total = total.head(100)
@@ -1306,12 +1087,7 @@ def q21(
     return total
 
 
-def q22(
-    root: str,
-    storage_options: Dict,
-    use_arrow_dtype: bool = None,
-    gpu: bool = False,
-):
+def q22(root: str, storage_options: Dict):
     customer = load_customer(root, storage_options)
     orders = load_orders(root, storage_options)
 
@@ -1342,20 +1118,13 @@ def q22(
         customer_filtered, on="C_CUSTKEY", how="inner"
     )
     customer_selected = customer_selected.loc[:, ["CNTRYCODE", "C_ACCTBAL"]]
-    agg1 = customer_selected.groupby(["CNTRYCODE"], as_index=False, sort=False).size()
+    agg1 = customer_selected.groupby(["CNTRYCODE"], as_index=False).size()
     agg1.columns = ["CNTRYCODE", "NUMCUST"]
-    agg2 = customer_selected.groupby(["CNTRYCODE"], as_index=False, sort=False).agg(
+    agg2 = customer_selected.groupby(["CNTRYCODE"], as_index=False).agg(
         TOTACCTBAL=pd.NamedAgg(column="C_ACCTBAL", aggfunc="sum")
     )
     total = agg1.merge(agg2, on="CNTRYCODE", how="inner")
-    total = total.sort_values(
-        by=[
-            "CNTRYCODE",
-        ],
-        ascending=[
-            True,
-        ],
-    )
+    total = total.sort_values(by=["CNTRYCODE"], ascending=[True])
 
     return total
 
@@ -1439,55 +1208,38 @@ def run_queries(
     queries,
     log_time=True,
     print_result=False,
-    use_arrow_dtype=False,
-    gpu=False,
+    include_io=False,
 ):
-    print("Start data loading")
-    total_start = time.time()
+    version = modin.__version__
+    data_start_time = time.time()
     for query in queries:
         loaders = query_to_loaders[query]
         for loader in loaders:
-            loader(
-                path,
-                storage_options=storage_options,
-                use_arrow_dtype=use_arrow_dtype,
-                gpu=gpu,
-            )
-    print(f"Data loading time (s): {time.time() - total_start}")
+            loader(path, storage_options)
+    print(f"Total data loading time (s): {time.time() - data_start_time}")
+
     total_start = time.time()
     for query in queries:
         try:
-            t1 = time.time()
-            result = query_to_runner[query](
-                path,
-                storage_options,
-                use_arrow_dtype=use_arrow_dtype,
-                gpu=gpu,
-            )
-            result = result.execute()
-            dur = time.time() - t1
+            start_time = time.time()
+            result = query_to_runner[query](path, storage_options)
+            without_io_time = time.time() - start_time
             success = True
             if print_result:
-                result_prefix = "./results"
-                if not os.path.exists(result_prefix):
-                    os.makedirs(result_prefix)
-                result_path = f"{result_prefix}/{query}.out"
-                result.to_csv(result_path, index=False)
+                print_result_fn("modin_ray", result, query)
         except Exception as e:
             print("".join(traceback.TracebackException.from_exception(e).format()))
-            dur = 0.0
+            without_io_time = 0.0
             success = False
         finally:
-            if log_time:
-                append_row("xorbits", query, dur, xorbits.__version__, success)
+            pass
+        if log_time:
+            log_time_fn("modin_ray", query, version=version, without_io_time=without_io_time, success=success)
     print(f"Total query execution time (s): {time.time() - total_start}")
 
 
 def main():
     parser = argparse.ArgumentParser(description="TPC-H benchmark.")
-    parser.add_argument(
-        "--path", type=str, required=True, help="path to the TPC-H dataset."
-    )
     parser.add_argument(
         "--storage_options",
         type=str,
@@ -1495,62 +1247,13 @@ def main():
         help="storage options json file.",
     )
     parser.add_argument(
-        "--queries",
-        type=int,
-        nargs="+",
-        required=False,
-        help="whitespace separated TPC-H queries to run.",
-    )
-    parser.add_argument(
         "--endpoint",
         type=str,
         required=False,
-        help="the endpoint of existing Xorbits cluster.",
+        help="the endpoint of existing Ray cluster.",
     )
-    parser.add_argument(
-        "--log_time",
-        action="store_true",
-        help="log time metrics or not.",
-    )
-    parser.add_argument(
-        "--print_result",
-        action="store_true",
-        help="print result.",
-    )
-    parser.add_argument(
-        "--use_arrow_dtype",
-        default=False,
-        action="store_true",
-        help="Use arrow dtype.",
-    )
-    parser.add_argument(
-        "--gpu",
-        default=False,
-        action="store_true",
-        help="Use GPUs.",
-    )
-    parser.add_argument(
-        "--cuda_devices",
-        type=int,
-        nargs="+",
-        required=False,
-        help="Comma separated cuda devices to use.",
-    )
-    parser.add_argument(
-        "--mmap_root_dir",
-        type=str,
-        required=False,
-        help="The directory used by mmap storage backend.",
-    )
-
+    parser = parse_common_arguments(parser)
     args = parser.parse_args()
-    print(f"Use GPU: {args.gpu}")
-    print(f"Use Arrow: {args.use_arrow_dtype}")
-    if args.mmap_root_dir is not None:
-        storage_config = {"mmap": {"root_dirs": args.mmap_root_dir}}
-        print(f"Use mmap to run: {args.mmap_root_dir}")
-    else:
-        storage_config = None
 
     # path to TPC-H data in parquet.
     print(f"Path: {args.path}")
@@ -1566,25 +1269,10 @@ def main():
     if args.queries is not None:
         queries = args.queries
     print(f"Queries to run: {queries}")
+    print(f"Include IO: {args.include_io}")
 
-    if args.endpoint == "local" or args.endpoint is None:
-        address = None
-    elif args.endpoint:
-        address = args.endpoint
-    xorbits.init(
-        address=address, cuda_devices=args.cuda_devices, storage_config=storage_config
-    )
-
-    try:
-        run_queries(
-            args.path,
-            storage_options,
-            queries,
-            args.log_time,
-            args.print_result,
-        )
-    finally:
-        xorbits.shutdown()
+    ray.init(address="auto")
+    run_queries(args.path, storage_options, queries, args.log_time, args.print_result, args.include_io)
 
 
 if __name__ == "__main__":
