+++ tpch/pyspark_queries/pandas_queries.py
@@ -1,152 +1,122 @@
 import os
-import json
 import time
 import argparse
 import traceback
-from datetime import datetime
-from typing import Dict
 
 import pandas as pd
 
-import dask
-import dask.dataframe as dd
-from dask.distributed import Client, wait
+import pyspark
+import pyspark.pandas as ps
+from pyspark.sql import SparkSession
 
-TIMINGS_FILE = "time.out"
+from common_utils import log_time_fn, parse_common_arguments, print_result_fn
+
 dataset_dict = {}
+spark: SparkSession = None
 
 
-def append_row(solution: str, q: str, secs: float, version: str, success=True):
-    with open(TIMINGS_FILE, "a") as f:
-        if f.tell() == 0:
-            f.write("solution,version,query_no,duration[s],success\n")
-        f.write(f"{solution},{version},{q},{secs},{success}\n")
-        print(f"{solution},{version},{q},{secs},{success}")
-
-
-def load_lineitem(root: str, storage_options: Dict):
+def load_lineitem(root: str):
     if "lineitem" not in dataset_dict:
         data_path = root + "/lineitem"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
-        df.L_SHIPDATE = dd.to_datetime(df.L_SHIPDATE, format="%Y-%m-%d")
-        df.L_RECEIPTDATE = dd.to_datetime(df.L_RECEIPTDATE, format="%Y-%m-%d")
-        df.L_COMMITDATE = dd.to_datetime(df.L_COMMITDATE, format="%Y-%m-%d")
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
+        print(df.count())
         dataset_dict["lineitem"] = df
-        result = df
     else:
-        result = dataset_dict["lineitem"]
-    return result
+        df = dataset_dict["lineitem"]
+    return df
 
 
-def load_part(root: str, storage_options: Dict):
-    if "part" not in dataset_dict or include_io:
+def load_part(root: str):
+    if "part" not in dataset_dict:
         data_path = root + "/part"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
+        print(df.count())
         dataset_dict["part"] = df
-        result = df
     else:
-        result = dataset_dict["part"]
-    return result
+        df = dataset_dict["part"]
+    return df
 
 
-def load_orders(root: str, storage_options: Dict):
-    if "orders" not in dataset_dict or include_io:
+def load_orders(root: str):
+    if "orders" not in dataset_dict:
         data_path = root + "/orders"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
-        df.O_ORDERDATE = dd.to_datetime(df.O_ORDERDATE, format="%Y-%m-%d")
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
+        print(df.count())
         dataset_dict["orders"] = df
-        result = df
     else:
-        result = dataset_dict["orders"]
-    return result
+        df = dataset_dict["orders"]
+    return df
 
 
-def load_customer(root: str, storage_options: Dict):
-    if "customer" not in dataset_dict or include_io:
+def load_customer(root: str):
+    if "customer" not in dataset_dict:
         data_path = root + "/customer"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
+        print(df.count())
         dataset_dict["customer"] = df
-        result = df
     else:
-        result = dataset_dict["customer"]
-    return result
+        df = dataset_dict["customer"]
+    return df
 
 
-def load_nation(root: str, storage_options: Dict):
-    if "nation" not in dataset_dict or include_io:
+def load_nation(root: str):
+    if "nation" not in dataset_dict:
         data_path = root + "/nation"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
+        print(df.count())
         dataset_dict["nation"] = df
-        result = df
     else:
-        result = dataset_dict["nation"]
-    return result
+        df = dataset_dict["nation"]
+    return df
 
 
-def load_region(root: str, storage_options: Dict):
-    if "region" not in dataset_dict or include_io:
+def load_region(root: str):
+    if "region" not in dataset_dict:
         data_path = root + "/region"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
+        print(df.count())
         dataset_dict["region"] = df
-        result = df
     else:
-        result = dataset_dict["region"]
-    return result
+        df = dataset_dict["region"]
+    return df
 
 
-def load_supplier(root: str, storage_options: Dict):
-    if "supplier" not in dataset_dict or include_io:
+def load_supplier(root: str):
+    if "supplier" not in dataset_dict:
         data_path = root + "/supplier"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
         dataset_dict["supplier"] = df
-        result = df
     else:
-        result = dataset_dict["supplier"]
-    return result
+        df = dataset_dict["supplier"]
+    return df
 
 
-def load_partsupp(root: str, storage_options: Dict):
-    if "partsupp" not in dataset_dict or include_io:
+def load_partsupp(root: str):
+    if "partsupp" not in dataset_dict:
         data_path = root + "/partsupp"
-        df = dd.read_parquet(
-            data_path,
-            storage_options=storage_options,
-        )
+        df = spark.read.parquet(data_path).cache()
+        df = ps.DataFrame(df)
         dataset_dict["partsupp"] = df
-        result = df
     else:
-        result = dataset_dict["partsupp"]
-    return result
+        df = dataset_dict["partsupp"]
+    return df
 
 
-def q01(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
+def q01(root: str):
+    lineitem = load_lineitem(root)
 
-    date = datetime.strptime("1998-09-02", "%Y-%m-%d")
+    date = pd.Timestamp("1998-09-02")
     lineitem_filtered = lineitem.loc[
         :,
         [
+            "L_ORDERKEY",
             "L_QUANTITY",
             "L_EXTENDEDPRICE",
             "L_DISCOUNT",
@@ -154,11 +124,10 @@ def q01(root: str, storage_options: Dict):
             "L_RETURNFLAG",
             "L_LINESTATUS",
             "L_SHIPDATE",
-            "L_ORDERKEY",
         ],
     ]
     sel = lineitem_filtered.L_SHIPDATE <= date
-    lineitem_filtered = lineitem_filtered[sel].copy()
+    lineitem_filtered = lineitem_filtered[sel]
     lineitem_filtered["AVG_QTY"] = lineitem_filtered.L_QUANTITY
     lineitem_filtered["AVG_PRICE"] = lineitem_filtered.L_EXTENDEDPRICE
     lineitem_filtered["DISC_PRICE"] = lineitem_filtered.L_EXTENDEDPRICE * (
@@ -169,8 +138,7 @@ def q01(root: str, storage_options: Dict):
         * (1 - lineitem_filtered.L_DISCOUNT)
         * (1 + lineitem_filtered.L_TAX)
     )
-    gb = lineitem_filtered.groupby(["L_RETURNFLAG", "L_LINESTATUS"])
-
+    gb = lineitem_filtered.groupby(["L_RETURNFLAG", "L_LINESTATUS"], as_index=False)
     total = gb.agg(
         {
             "L_QUANTITY": "sum",
@@ -183,20 +151,35 @@ def q01(root: str, storage_options: Dict):
             "L_ORDERKEY": "count",
         }
     )
+    total = (
+        total.reset_index()
+            .sort_values(["L_RETURNFLAG", "L_LINESTATUS"])
+            .rename(columns={
+                "L_QUANTITY": "SUM_QTY",
+                "L_EXTENDEDPRICE": "SUM_BASE_PRICE",
+                "DISC_PRICE": "SUM_DISC_PRICE",
+                "CHARGE": "SUM_CHARGE",
+                "L_DISCOUNT": "AVG_DISC",
+                "L_ORDERKEY": "COUNT_ORDER"
+            })
+    )
 
-    total = total.compute().reset_index().sort_values(["L_RETURNFLAG", "L_LINESTATUS"])
     return total
 
 
-def q02(root: str, storage_options: Dict):
-    part = load_part(root, storage_options)
-    partsupp = load_partsupp(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
-    region = load_region(root, storage_options)
+def q02(root: str):
+    nation = load_nation(root)
+    region = load_region(root)
+    supplier = load_supplier(root)
+    part = load_part(root)
+    partsupp = load_partsupp(root)
+
+    size = 15
+    p_type = "BRASS"
+    region_name = "EUROPE"
 
     nation_filtered = nation.loc[:, ["N_NATIONKEY", "N_NAME", "N_REGIONKEY"]]
-    region_filtered = region[(region["R_NAME"] == "EUROPE")]
+    region_filtered = region[(region["R_NAME"] == region_name)]
     region_filtered = region_filtered.loc[:, ["R_REGIONKEY"]]
     r_n_merged = nation_filtered.merge(
         region_filtered, left_on="N_REGIONKEY", right_on="R_REGIONKEY", how="inner"
@@ -248,8 +231,8 @@ def q02(root: str, storage_options: Dict):
     ]
     part_filtered = part.loc[:, ["P_PARTKEY", "P_MFGR", "P_SIZE", "P_TYPE"]]
     part_filtered = part_filtered[
-        (part_filtered["P_SIZE"] == 15)
-        & (part_filtered["P_TYPE"].str.endswith("BRASS"))
+        (part_filtered["P_SIZE"] == size)
+        & (part_filtered["P_TYPE"].str.endswith(p_type))
     ]
     part_filtered = part_filtered.loc[:, ["P_PARTKEY", "P_MFGR"]]
     merged_df = part_filtered.merge(
@@ -269,9 +252,7 @@ def q02(root: str, storage_options: Dict):
             "P_MFGR",
         ],
     ]
-
-    min_values = merged_df.groupby("P_PARTKEY")["PS_SUPPLYCOST"].min().reset_index()
-
+    min_values = merged_df.groupby("P_PARTKEY", as_index=False)["PS_SUPPLYCOST"].min()
     min_values.columns = ["P_PARTKEY_CPY", "MIN_SUPPLYCOST"]
     merged_df = merged_df.merge(
         min_values,
@@ -292,98 +273,86 @@ def q02(root: str, storage_options: Dict):
             "S_COMMENT",
         ],
     ]
-
-    total = total.compute().sort_values(
-        by=[
-            "S_ACCTBAL",
-            "N_NAME",
-            "S_NAME",
-            "P_PARTKEY",
-        ],
-        ascending=[
-            False,
-            True,
-            True,
-            True,
-        ],
+    total = total.sort_values(
+        by=["S_ACCTBAL", "N_NAME", "S_NAME", "P_PARTKEY"],
+        ascending=[False, True, True, True],
     )
-
     total = total.head(100)
 
     return total
 
 
-def q03(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    customer = load_customer(root, storage_options)
-
-    date = datetime.strptime("1995-03-04", "%Y-%m-%d")
+def q03(root: str):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    customer = load_customer(root)
+    mktsegment = "HOUSEHOLD"
+    date = pd.Timestamp("1995-03-04")
     lineitem_filtered = lineitem.loc[
         :, ["L_ORDERKEY", "L_EXTENDEDPRICE", "L_DISCOUNT", "L_SHIPDATE"]
     ]
-    orders_filtered = orders.loc[
-        :, ["O_ORDERKEY", "O_CUSTKEY", "O_ORDERDATE", "O_SHIPPRIORITY"]
-    ]
+    orders_filtered = orders.loc[:, ["O_ORDERKEY", "O_CUSTKEY", "O_ORDERDATE", "O_SHIPPRIORITY"]]
     customer_filtered = customer.loc[:, ["C_MKTSEGMENT", "C_CUSTKEY"]]
     lsel = lineitem_filtered.L_SHIPDATE > date
     osel = orders_filtered.O_ORDERDATE < date
-    csel = customer_filtered.C_MKTSEGMENT == "HOUSEHOLD"
+    csel = customer_filtered.C_MKTSEGMENT == mktsegment
     flineitem = lineitem_filtered[lsel]
     forders = orders_filtered[osel]
     fcustomer = customer_filtered[csel]
     jn1 = fcustomer.merge(forders, left_on="C_CUSTKEY", right_on="O_CUSTKEY")
     jn2 = jn1.merge(flineitem, left_on="O_ORDERKEY", right_on="L_ORDERKEY")
-    jn2["TMP"] = jn2.L_EXTENDEDPRICE * (1 - jn2.L_DISCOUNT)
+    jn2["REVENUE"] = jn2.L_EXTENDEDPRICE * (1 - jn2.L_DISCOUNT)
     total = (
-        jn2.groupby(["L_ORDERKEY", "O_ORDERDATE", "O_SHIPPRIORITY"])["TMP"]
+        jn2.groupby(["L_ORDERKEY", "O_ORDERDATE", "O_SHIPPRIORITY"], as_index=False)["REVENUE"]
         .sum()
-        .compute()
-        .reset_index()
-        .sort_values(["TMP"], ascending=False)
+        .sort_values(["REVENUE"], ascending=False)
     )
 
-    res = total.loc[:, ["L_ORDERKEY", "TMP", "O_ORDERDATE", "O_SHIPPRIORITY"]]
-    res = res.head(10)
-    return res
+    total = total[:10].loc[
+        :, ["L_ORDERKEY", "REVENUE", "O_ORDERDATE", "O_SHIPPRIORITY"]
+    ]
+
+    return total
 
 
-def q04(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
+def q04(root: str):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
 
-    date1 = datetime.strptime("1993-11-01", "%Y-%m-%d")
-    date2 = datetime.strptime("1993-08-01", "%Y-%m-%d")
+    date2 = pd.Timestamp("1993-8-01")
+    date1 = date2 + pd.DateOffset(months=3)
+    date2 = pd.Timestamp("1993-08-01")
     lsel = lineitem.L_COMMITDATE < lineitem.L_RECEIPTDATE
     osel = (orders.O_ORDERDATE < date1) & (orders.O_ORDERDATE >= date2)
     flineitem = lineitem[lsel]
     forders = orders[osel]
+    # pandas API on Spark does not support isin
+    # jn = forders[forders["O_ORDERKEY"].isin(flineitem["L_ORDERKEY"])]
     forders = forders[["O_ORDERKEY", "O_ORDERPRIORITY"]]
-    jn = forders.merge(
-        flineitem, left_on="O_ORDERKEY", right_on="L_ORDERKEY"
-    ).drop_duplicates(subset=["O_ORDERKEY"])[["O_ORDERPRIORITY", "O_ORDERKEY"]]
+    jn = forders.merge(flineitem, left_on="O_ORDERKEY", right_on="L_ORDERKEY") \
+            .drop_duplicates(subset=["O_ORDERKEY"])[["O_ORDERPRIORITY", "O_ORDERKEY"]]
     total = (
-        jn.groupby("O_ORDERPRIORITY")["O_ORDERKEY"]
+        jn.groupby("O_ORDERPRIORITY", as_index=False)["O_ORDERKEY"]
         .count()
-        .reset_index()
         .sort_values(["O_ORDERPRIORITY"])
+        .rename(columns={"O_ORDERKEY": "ORDER_COUNT"})
     )
-    total = total.compute()
+
     return total
 
 
-def q05(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    customer = load_customer(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
-    region = load_region(root, storage_options)
+def q05(root):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    customer = load_customer(root)
+    supplier = load_supplier(root)
+    nation = load_nation(root)
+    region = load_region(root)
 
-    date1 = datetime.strptime("1996-01-01", "%Y-%m-%d")
-    date2 = datetime.strptime("1997-01-01", "%Y-%m-%d")
-
-    rsel = region.R_NAME == "ASIA"
+    region_name = "ASIA"
+    date1 = pd.Timestamp("1996-01-01")
+    date2 = date1 + pd.DateOffset(years=1)
+    rsel = region.R_NAME == region_name
     osel = (orders.O_ORDERDATE >= date1) & (orders.O_ORDERDATE < date2)
 
     forders = orders[osel]
@@ -396,19 +365,18 @@ def q05(root: str, storage_options: Dict):
     jn5 = supplier.merge(
         jn4, left_on=["S_SUPPKEY", "S_NATIONKEY"], right_on=["L_SUPPKEY", "N_NATIONKEY"]
     )
-    jn5["TMP"] = jn5.L_EXTENDEDPRICE * (1.0 - jn5.L_DISCOUNT)
-    gb = jn5.groupby("N_NAME")["TMP"].sum()
+    jn5["REVENUE"] = jn5.L_EXTENDEDPRICE * (1.0 - jn5.L_DISCOUNT)
+    gb = jn5.groupby("N_NAME", as_index=False)["REVENUE"].sum()
+    total = gb.sort_values("REVENUE", ascending=False)
 
-    total = gb.compute().reset_index().sort_values("TMP", ascending=False)
     return total
 
 
-def q06(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
+def q06(root: str):
+    lineitem = load_lineitem(root)
 
-    date1 = datetime.strptime("1996-01-01", "%Y-%m-%d")
-    date2 = datetime.strptime("1997-01-01", "%Y-%m-%d")
-
+    date1 = pd.Timestamp("1996-01-01")
+    date2 = date1 + pd.DateOffset(years=1)
     lineitem_filtered = lineitem.loc[
         :, ["L_QUANTITY", "L_EXTENDEDPRICE", "L_DISCOUNT", "L_SHIPDATE"]
     ]
@@ -420,25 +388,26 @@ def q06(root: str, storage_options: Dict):
         & (lineitem_filtered.L_QUANTITY < 24)
     )
     flineitem = lineitem_filtered[sel]
-    total = (flineitem.L_EXTENDEDPRICE * flineitem.L_DISCOUNT).sum()
-    total = total.compute()
-    return total
+    result_value = (flineitem.L_EXTENDEDPRICE * flineitem.L_DISCOUNT).sum()
+    result_df = ps.DataFrame({"REVENUE": [result_value]})
+
+    return result_df
 
 
-def q07(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    customer = load_customer(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
+def q07(root: str):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    customer = load_customer(root)
+    supplier = load_supplier(root)
+    nation = load_nation(root)
 
+    nation1 = "FRANCE"
+    nation2 = "GERMANY"
     lineitem_filtered = lineitem[
-        (lineitem["L_SHIPDATE"] >= datetime.strptime("1995-01-01", "%Y-%m-%d"))
-        & (lineitem["L_SHIPDATE"] < datetime.strptime("1997-01-01", "%Y-%m-%d"))
+        (lineitem["L_SHIPDATE"] >= pd.Timestamp("1995-01-01"))
+        & (lineitem["L_SHIPDATE"] < pd.Timestamp("1997-01-01"))
     ]
-    lineitem_filtered["L_YEAR"] = lineitem_filtered["L_SHIPDATE"].apply(
-        lambda x: x.year
-    )
+    lineitem_filtered["L_YEAR"] = lineitem_filtered["L_SHIPDATE"].dt.year
     lineitem_filtered["VOLUME"] = lineitem_filtered["L_EXTENDEDPRICE"] * (
         1.0 - lineitem_filtered["L_DISCOUNT"]
     )
@@ -448,8 +417,8 @@ def q07(root: str, storage_options: Dict):
     supplier_filtered = supplier.loc[:, ["S_SUPPKEY", "S_NATIONKEY"]]
     orders_filtered = orders.loc[:, ["O_ORDERKEY", "O_CUSTKEY"]]
     customer_filtered = customer.loc[:, ["C_CUSTKEY", "C_NATIONKEY"]]
-    n1 = nation[(nation["N_NAME"] == "FRANCE")].loc[:, ["N_NATIONKEY", "N_NAME"]]
-    n2 = nation[(nation["N_NAME"] == "GERMANY")].loc[:, ["N_NATIONKEY", "N_NAME"]]
+    n1 = nation[(nation["N_NAME"] == nation1)].loc[:, ["N_NATIONKEY", "N_NAME"]]
+    n2 = nation[(nation["N_NAME"] == nation2)].loc[:, ["N_NATIONKEY", "N_NAME"]]
 
     # ----- do nation 1 -----
     N1_C = customer_filtered.merge(
@@ -508,40 +477,43 @@ def q07(root: str, storage_options: Dict):
     total2 = total2.drop(columns=["O_ORDERKEY", "L_ORDERKEY"])
 
     # concat results
-    total = dd.concat([total1, total2])
-    total = total.groupby(["SUPP_NATION", "CUST_NATION", "L_YEAR"]).VOLUME.agg("sum")
+    total = ps.concat([total1, total2])
+    
+    # pandas API on spark does not support groupby.agg(COL=pd.NamedAgg)
+    # total = (
+    #     total.groupby(["SUPP_NATION", "CUST_NATION", "L_YEAR"], as_index=False)
+    #         .agg(REVENUE=ps.NamedAgg(column="VOLUME", aggfunc="sum"))
+    #         .sort_values(
+    #             by=["SUPP_NATION", "CUST_NATION", "L_YEAR"],
+    #             ascending=[True, True, True]
+    #         )
+    # )
+    total = total.groupby(["SUPP_NATION", "CUST_NATION", "L_YEAR"], as_index=False)["VOLUME"].sum()
     total.columns = ["SUPP_NATION", "CUST_NATION", "L_YEAR", "REVENUE"]
+    total = total.sort_values(
+                by=["SUPP_NATION", "CUST_NATION", "L_YEAR"],
+                ascending=[True, True, True]
+            )
 
-    total = (
-        total.compute()
-        .reset_index()
-        .sort_values(
-            by=["SUPP_NATION", "CUST_NATION", "L_YEAR"],
-            ascending=[
-                True,
-                True,
-                True,
-            ],
-        )
-    )
     return total
 
 
-def q08(root: str, storage_options: Dict):
-    part = load_part(root, storage_options)
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    customer = load_customer(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
-    region = load_region(root, storage_options)
+def q08(root):
+    part = load_part(root)
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    customer = load_customer(root)
+    supplier = load_supplier(root)
+    nation = load_nation(root)
+    region = load_region(root)
 
-    part_filtered = part[(part["P_TYPE"] == "ECONOMY ANODIZED STEEL")]
+    nation_name = "BRAZIL"
+    region_name = "AMERICA"
+    p_type = "ECONOMY ANODIZED STEEL"
+    part_filtered = part[(part["P_TYPE"] == p_type)]
     part_filtered = part_filtered.loc[:, ["P_PARTKEY"]]
     lineitem_filtered = lineitem.loc[:, ["L_PARTKEY", "L_SUPPKEY", "L_ORDERKEY"]]
-    lineitem_filtered["VOLUME"] = lineitem["L_EXTENDEDPRICE"] * (
-        1.0 - lineitem["L_DISCOUNT"]
-    )
+    lineitem_filtered["VOLUME"] = lineitem["L_EXTENDEDPRICE"] * (1.0 - lineitem["L_DISCOUNT"])
     total = part_filtered.merge(
         lineitem_filtered, left_on="P_PARTKEY", right_on="L_PARTKEY", how="inner"
     )
@@ -552,10 +524,10 @@ def q08(root: str, storage_options: Dict):
     )
     total = total.loc[:, ["L_ORDERKEY", "VOLUME", "S_NATIONKEY"]]
     orders_filtered = orders[
-        (orders["O_ORDERDATE"] >= datetime.strptime("1995-01-01", "%Y-%m-%d"))
-        & (orders["O_ORDERDATE"] < datetime.strptime("1997-01-01", "%Y-%m-%d"))
+        (orders["O_ORDERDATE"] >= pd.Timestamp("1995-01-01"))
+        & (orders["O_ORDERDATE"] < pd.Timestamp("1997-01-01"))
     ]
-    orders_filtered["O_YEAR"] = orders_filtered["O_ORDERDATE"].apply(lambda x: x.year)
+    orders_filtered["O_YEAR"] = orders_filtered["O_ORDERDATE"].dt.year
     orders_filtered = orders_filtered.loc[:, ["O_ORDERKEY", "O_CUSTKEY", "O_YEAR"]]
     total = total.merge(
         orders_filtered, left_on="L_ORDERKEY", right_on="O_ORDERKEY", how="inner"
@@ -567,9 +539,8 @@ def q08(root: str, storage_options: Dict):
     )
     total = total.loc[:, ["VOLUME", "S_NATIONKEY", "O_YEAR", "C_NATIONKEY"]]
     n1_filtered = nation.loc[:, ["N_NATIONKEY", "N_REGIONKEY"]]
-    n2_filtered = nation.loc[:, ["N_NATIONKEY", "N_NAME"]].rename(
-        columns={"N_NAME": "NATION"}
-    )
+    n2_filtered = nation.loc[:, ["N_NATIONKEY", "N_NAME"]] \
+        .rename(columns={"N_NAME": "NATION"})
     total = total.merge(
         n1_filtered, left_on="C_NATIONKEY", right_on="N_NATIONKEY", how="inner"
     )
@@ -578,7 +549,7 @@ def q08(root: str, storage_options: Dict):
         n2_filtered, left_on="S_NATIONKEY", right_on="N_NATIONKEY", how="inner"
     )
     total = total.loc[:, ["VOLUME", "O_YEAR", "N_REGIONKEY", "NATION"]]
-    region_filtered = region[(region["R_NAME"] == "AMERICA")]
+    region_filtered = region[(region["R_NAME"] == region_name)]
     region_filtered = region_filtered.loc[:, ["R_REGIONKEY"]]
     total = total.merge(
         region_filtered, left_on="N_REGIONKEY", right_on="R_REGIONKEY", how="inner"
@@ -587,36 +558,27 @@ def q08(root: str, storage_options: Dict):
 
     def udf(df):
         demonimator = df["VOLUME"].sum()
-        df = df[df["NATION"] == "BRAZIL"]
+        df = df[df["NATION"] == nation_name]
         numerator = df["VOLUME"].sum()
         return numerator / demonimator
 
-    total = total.groupby("O_YEAR").apply(udf)
-    total = (
-        total.compute()
-        .reset_index()
-        .sort_values(
-            by=[
-                "O_YEAR",
-            ],
-            ascending=[
-                True,
-            ],
-        )
-    )
+    total = total.groupby("O_YEAR").apply(udf).reset_index()
     total.columns = ["O_YEAR", "MKT_SHARE"]
+    total = total.sort_values(by=["O_YEAR"], ascending=[True])
+
     return total
 
 
-def q09(root: str, storage_options: Dict):
-    part = load_part(root, storage_options)
-    partsupp = load_partsupp(root, storage_options)
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
+def q09(root: str):
+    part = load_part(root)
+    partsupp = load_partsupp(root)
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    supplier = load_supplier(root)
+    nation = load_nation(root)
 
-    psel = part.P_NAME.str.contains("ghost")
+    p_name = "ghost"
+    psel = part.P_NAME.str.contains(p_name)
     fpart = part[psel]
     jn1 = lineitem.merge(fpart, left_on="L_PARTKEY", right_on="P_PARTKEY")
     jn2 = jn1.merge(supplier, left_on="L_SUPPKEY", right_on="S_SUPPKEY")
@@ -628,25 +590,22 @@ def q09(root: str, storage_options: Dict):
     jn5["TMP"] = jn5.L_EXTENDEDPRICE * (1 - jn5.L_DISCOUNT) - (
         (1 * jn5.PS_SUPPLYCOST) * jn5.L_QUANTITY
     )
-    jn5["O_YEAR"] = jn5.O_ORDERDATE.apply(lambda x: x.year)
-    gb = jn5.groupby(["N_NAME", "O_YEAR"])["TMP"].sum()
-    total = (
-        gb.compute()
-        .reset_index()
-        .sort_values(["N_NAME", "O_YEAR"], ascending=[True, False])
-    )
+    jn5["O_YEAR"] = jn5.O_ORDERDATE.dt.year
+    gb = jn5.groupby(["N_NAME", "O_YEAR"], as_index=False)["TMP"].sum()
+    total = gb.sort_values(["N_NAME", "O_YEAR"], ascending=[True, False])
+    total = total.rename(columns={"TMP": "SUM_PROFIT"})
+
     return total
 
 
-def q10(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    nation = load_nation(root, storage_options)
-    customer = load_customer(root, storage_options)
+def q10(root: str):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    nation = load_nation(root)
+    customer = load_customer(root)
 
-    t1 = time.time()
-    date1 = datetime.strptime("1994-11-01", "%Y-%m-%d")
-    date2 = datetime.strptime("1995-02-01", "%Y-%m-%d")
+    date1 = pd.Timestamp("1994-11-01")
+    date2 = date1 + pd.DateOffset(months=3)
     osel = (orders.O_ORDERDATE >= date1) & (orders.O_ORDERDATE < date2)
     lsel = lineitem.L_RETURNFLAG == "R"
     forders = orders[osel]
@@ -654,7 +613,7 @@ def q10(root: str, storage_options: Dict):
     jn1 = flineitem.merge(forders, left_on="L_ORDERKEY", right_on="O_ORDERKEY")
     jn2 = jn1.merge(customer, left_on="O_CUSTKEY", right_on="C_CUSTKEY")
     jn3 = jn2.merge(nation, left_on="C_NATIONKEY", right_on="N_NATIONKEY")
-    jn3["TMP"] = jn3.L_EXTENDEDPRICE * (1.0 - jn3.L_DISCOUNT)
+    jn3["REVENUE"] = jn3.L_EXTENDEDPRICE * (1.0 - jn3.L_DISCOUNT)
     gb = jn3.groupby(
         [
             "C_CUSTKEY",
@@ -665,17 +624,33 @@ def q10(root: str, storage_options: Dict):
             "C_ADDRESS",
             "C_COMMENT",
         ],
-    )["TMP"].sum()
-    total = gb.compute().reset_index().sort_values("TMP", ascending=False)
+        as_index=False,
+    )["REVENUE"].sum()
+    total = gb.sort_values("REVENUE", ascending=False)
     total = total.head(20)
+    total = total[
+        [
+            "C_CUSTKEY",
+            "C_NAME",
+            "REVENUE",
+            "C_ACCTBAL",
+            "N_NAME",
+            "C_ADDRESS",
+            "C_PHONE",
+            "C_COMMENT",
+        ]
+    ]
 
     return total
 
 
-def q11(root: str, storage_options: Dict):
-    partsupp = load_partsupp(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
+def q11(root: str):
+    partsupp = load_partsupp(root)
+    supplier = load_supplier(root)
+    nation = load_nation(root)
+
+    nation_name = "GERMANY"
+    fraction = 0.0001
 
     partsupp_filtered = partsupp.loc[:, ["PS_PARTKEY", "PS_SUPPKEY"]]
     partsupp_filtered["TOTAL_COST"] = (
@@ -685,29 +660,31 @@ def q11(root: str, storage_options: Dict):
     ps_supp_merge = partsupp_filtered.merge(
         supplier_filtered, left_on="PS_SUPPKEY", right_on="S_SUPPKEY", how="inner"
     )
-    ps_supp_merge.loc[:, ["PS_PARTKEY", "S_NATIONKEY", "TOTAL_COST"]]
-    nation_filtered = nation[(nation["N_NAME"] == "GERMANY")]
+    ps_supp_merge = ps_supp_merge.loc[:, ["PS_PARTKEY", "S_NATIONKEY", "TOTAL_COST"]]
+    nation_filtered = nation[(nation["N_NAME"] == nation_name)]
     nation_filtered = nation_filtered.loc[:, ["N_NATIONKEY"]]
     ps_supp_n_merge = ps_supp_merge.merge(
         nation_filtered, left_on="S_NATIONKEY", right_on="N_NATIONKEY", how="inner"
     )
     ps_supp_n_merge = ps_supp_n_merge.loc[:, ["PS_PARTKEY", "TOTAL_COST"]]
-    sum_val = ps_supp_n_merge["TOTAL_COST"].sum() * 0.0001
-
-    total = ps_supp_n_merge.groupby(["PS_PARTKEY"]).TOTAL_COST.agg("sum").reset_index()
-    total = total.rename(columns={"TOTAL_COST": "VALUE"})
+    sum_val = ps_supp_n_merge["TOTAL_COST"].sum() * fraction
+    total = ps_supp_n_merge.groupby(["PS_PARTKEY"], as_index=False).agg(
+        VALUE=ps.NamedAgg(column="TOTAL_COST", aggfunc="sum")
+    )
     total = total[total["VALUE"] > sum_val]
-    total = total.compute().sort_values("VALUE", ascending=False)
-    
+    total = total.sort_values("VALUE", ascending=False)
+
     return total
 
 
-def q12(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
+def q12(root):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
 
-    date1 = datetime.strptime("1994-01-01", "%Y-%m-%d")
-    date2 = datetime.strptime("1995-01-01", "%Y-%m-%d")
+    shipmode1 = "MAIL"
+    shipmode2 = "SHIP"
+    date1 = pd.Timestamp("1994-01-01")
+    date2 = date1 + pd.DateOffset(years=1)
     sel = (
         (lineitem.L_RECEIPTDATE < date2)
         & (lineitem.L_COMMITDATE < date2)
@@ -715,95 +692,103 @@ def q12(root: str, storage_options: Dict):
         & (lineitem.L_SHIPDATE < lineitem.L_COMMITDATE)
         & (lineitem.L_COMMITDATE < lineitem.L_RECEIPTDATE)
         & (lineitem.L_RECEIPTDATE >= date1)
-        & ((lineitem.L_SHIPMODE == "MAIL") | (lineitem.L_SHIPMODE == "SHIP"))
+        & ((lineitem.L_SHIPMODE == shipmode1) | (lineitem.L_SHIPMODE == shipmode2))
     )
     flineitem = lineitem[sel]
     jn = flineitem.merge(orders, left_on="L_ORDERKEY", right_on="O_ORDERKEY")
-    gb = jn.groupby("L_SHIPMODE")["O_ORDERPRIORITY"]
 
-    def g1(x):
-        return x.apply(lambda s: ((s == "1-URGENT") | (s == "2-HIGH")).sum())
+    # the UDFs of pandas API on Spark are different from pandas
+    # def g1(x):
+    #     return ((x == "1-URGENT") | (x == "2-HIGH")).sum()
 
-    def g2(x):
-        return x.apply(lambda s: ((s != "1-URGENT") & (s != "2-HIGH")).sum())
+    # def g2(x):
+    #     return ((x != "1-URGENT") & (x != "2-HIGH")).sum()
+    
+    # total = jn.groupby("L_SHIPMODE", as_index=False)["O_ORDERPRIORITY"] \
+    #     .agg({"O_ORDERPRIORITY": [g1, g2]})
+    # total = (
+    #     total.sort_values("L_SHIPMODE")
+    #     .rename(columns={"g1": "HIGH_LINE_COUNT", "g2": "LOW_LINE_COUNT"})
+    # )
 
-    g1_agg = dd.Aggregation("g1", g1, lambda s0: s0.sum())
-    g2_agg = dd.Aggregation("g2", g2, lambda s0: s0.sum())
-    total = gb.agg([g1_agg, g2_agg])
-    total = total.compute().reset_index().sort_values("L_SHIPMODE")
-
+    jn['HIGH_LINE_COUNT'] = jn['O_ORDERPRIORITY'].isin(["1-URGENT", "2-HIGH"]).astype('long')
+    jn['LOW_LINE_COUNT'] = (~jn['O_ORDERPRIORITY'].isin(["1-URGENT", "2-HIGH"])).astype('long')
+    total = jn.groupby("L_SHIPMODE") \
+        .agg({"HIGH_LINE_COUNT": "sum", 'LOW_LINE_COUNT': 'sum'}) \
+        .reset_index()
+    total = total.sort_values("L_SHIPMODE")
+    
     return total
 
 
-def q13(root: str, storage_options: Dict):
-    customer = load_customer(root, storage_options)
-    orders = load_orders(root, storage_options)
+def q13(root: str):
+    customer = load_customer(root)
+    orders = load_orders(root)
 
+    word1 = "special"
+    word2 = "requests"
     customer_filtered = customer.loc[:, ["C_CUSTKEY"]]
     orders_filtered = orders[
-        ~orders["O_COMMENT"].str.contains("special(\\S|\\s)*requests")
+        ~orders["O_COMMENT"].str.contains(f"{word1}(\\S|\\s)*{word2}")
     ]
     orders_filtered = orders_filtered.loc[:, ["O_ORDERKEY", "O_CUSTKEY"]]
     c_o_merged = customer_filtered.merge(
         orders_filtered, left_on="C_CUSTKEY", right_on="O_CUSTKEY", how="left"
     )
     c_o_merged = c_o_merged.loc[:, ["C_CUSTKEY", "O_ORDERKEY"]]
-
-    count_df = c_o_merged.groupby(["C_CUSTKEY"]).O_ORDERKEY.agg("count").reset_index()
-    count_df = count_df.rename(columns={"O_ORDERKEY": "C_COUNT"})
-
-    total = count_df.groupby(["C_COUNT"]).size().reset_index()
+    count_df = c_o_merged.groupby(["C_CUSTKEY"], as_index=False).agg(
+        C_COUNT=ps.NamedAgg(column="O_ORDERKEY", aggfunc="count")
+    )
+    total = count_df.groupby(["C_COUNT"], as_index=False).size().reset_index()
     total.columns = ["C_COUNT", "CUSTDIST"]
-    total = total.compute().sort_values(
+    total = total.sort_values(
         by=["CUSTDIST", "C_COUNT"],
-        ascending=[
-            False,
-            False,
-        ],
+        ascending=[False, False],
     )
 
     return total
 
 
-def q14(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    part = load_part(root, storage_options)
+def q14(root: str):
+    lineitem = load_lineitem(root)
+    part = load_part(root)
 
-    startDate = datetime.strptime("1994-03-01", "%Y-%m-%d")
-    endDate = datetime.strptime("1994-04-01", "%Y-%m-%d")
+    startDate = pd.Timestamp("1994-03-01")
+    endDate = startDate + pd.DateOffset(months=1)
     p_type_like = "PROMO"
     part_filtered = part.loc[:, ["P_PARTKEY", "P_TYPE"]]
     lineitem_filtered = lineitem.loc[
         :, ["L_EXTENDEDPRICE", "L_DISCOUNT", "L_SHIPDATE", "L_PARTKEY"]
     ]
-    sel = (lineitem_filtered.L_SHIPDATE >= startDate) & (
-        lineitem_filtered.L_SHIPDATE < endDate
-    )
+    sel = (lineitem_filtered.L_SHIPDATE >= startDate) \
+        & (lineitem_filtered.L_SHIPDATE < endDate)
     flineitem = lineitem_filtered[sel]
     jn = flineitem.merge(part_filtered, left_on="L_PARTKEY", right_on="P_PARTKEY")
-    jn["TMP"] = jn.L_EXTENDEDPRICE * (1.0 - jn.L_DISCOUNT)
-    total = jn[jn.P_TYPE.str.startswith(p_type_like)].TMP.sum() * 100 / jn.TMP.sum()
+    jn["PROMO_REVENUE"] = jn.L_EXTENDEDPRICE * (1.0 - jn.L_DISCOUNT)
+    total = (
+        jn[jn.P_TYPE.str.startswith(p_type_like)].PROMO_REVENUE.sum() * 100 / jn.PROMO_REVENUE.sum()
+    )
 
-    return total
+    result_df = ps.DataFrame({"PROMO_REVENUE": [total]})
+    return result_df
 
-def q15(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    supplier = load_supplier(root, storage_options)
+
+def q15(root: str):
+    lineitem = load_lineitem(root)
+    supplier = load_supplier(root)
 
     lineitem_filtered = lineitem[
-        (lineitem["L_SHIPDATE"] >= datetime.strptime("1996-01-01", "%Y-%m-%d"))
-        & (lineitem["L_SHIPDATE"] < (datetime.strptime("1996-04-01", "%Y-%m-%d")))
-    ]  # + pd.DateOffset(months=3)))]
+        (lineitem["L_SHIPDATE"] >= pd.Timestamp("1996-01-01"))
+        & (lineitem["L_SHIPDATE"] < (pd.Timestamp("1996-01-01") + pd.DateOffset(months=3)))
+    ]
     lineitem_filtered["REVENUE_PARTS"] = lineitem_filtered["L_EXTENDEDPRICE"] * (
         1.0 - lineitem_filtered["L_DISCOUNT"]
     )
     lineitem_filtered = lineitem_filtered.loc[:, ["L_SUPPKEY", "REVENUE_PARTS"]]
-    revenue_table = (
-        lineitem_filtered.groupby("L_SUPPKEY")["REVENUE_PARTS"].agg("sum").reset_index()
-    )
-    revenue_table = revenue_table.rename(
-        columns={"REVENUE_PARTS": "TOTAL_REVENUE", "L_SUPPKEY": "SUPPLIER_NO"}
-    )
+    revenue_table = lineitem_filtered.groupby("L_SUPPKEY", as_index=False).sum()
+    revenue_table.columns = ["L_SUPPKEY", "TOTAL_REVENUE"]
+    revenue_table = revenue_table.rename(columns={"L_SUPPKEY": "SUPPLIER_NO"})
+
     max_revenue = revenue_table["TOTAL_REVENUE"].max()
     revenue_table = revenue_table[revenue_table["TOTAL_REVENUE"] == max_revenue]
     supplier_filtered = supplier.loc[:, ["S_SUPPKEY", "S_NAME", "S_ADDRESS", "S_PHONE"]]
@@ -813,19 +798,29 @@ def q15(root: str, storage_options: Dict):
     total = total.loc[
         :, ["S_SUPPKEY", "S_NAME", "S_ADDRESS", "S_PHONE", "TOTAL_REVENUE"]
     ]
-    
+
     return total
 
 
-def q16(root: str, storage_options: Dict):
-    part = load_part(root, storage_options)
-    partsupp = load_partsupp(root, storage_options)
-    supplier = load_supplier(root, storage_options)
+def q16(root: str):
+    part = load_part(root)
+    partsupp = load_partsupp(root)
+    supplier = load_supplier(root)
 
+    brand = "Brand#45"
+    p_type = "MEDIUM POLISHED"
+    size1 = 49
+    size2 = 14
+    size3 = 23
+    size4 = 45
+    size5 = 19
+    size6 = 3
+    size7 = 36
+    size8 = 9
     part_filtered = part[
-        (part["P_BRAND"] != "Brand#45")
-        & (~part["P_TYPE"].str.contains("^MEDIUM POLISHED"))
-        & part["P_SIZE"].isin([49, 14, 23, 45, 19, 3, 36, 9])
+        (part["P_BRAND"] != brand)
+        & (~part["P_TYPE"].str.contains(f"^{p_type}"))
+        & part["P_SIZE"].isin([size1, size2, size3, size4, size5, size6, size7, size8])
     ]
     part_filtered = part_filtered.loc[:, ["P_PARTKEY", "P_BRAND", "P_TYPE", "P_SIZE"]]
     partsupp_filtered = partsupp.loc[:, ["PS_PARTKEY", "PS_SUPPKEY"]]
@@ -837,19 +832,16 @@ def q16(root: str, storage_options: Dict):
         supplier["S_COMMENT"].str.contains("Customer(\\S|\\s)*Complaints")
     ]
     supplier_filtered = supplier_filtered.loc[:, ["S_SUPPKEY"]].drop_duplicates()
-    # left merge to select only ps_suppkey values not in supplier_filtered
+    # left merge to select only PS_SUPPKEY values not in supplier_filtered
     total = total.merge(
         supplier_filtered, left_on="PS_SUPPKEY", right_on="S_SUPPKEY", how="left"
     )
     total = total[total["S_SUPPKEY"].isna()]
     total = total.loc[:, ["P_BRAND", "P_TYPE", "P_SIZE", "PS_SUPPKEY"]]
-    total = (
-        total.groupby(["P_BRAND", "P_TYPE", "P_SIZE"])["PS_SUPPKEY"]
-        .nunique()
-        .reset_index()
-    )
+    total = total.groupby(["P_BRAND", "P_TYPE", "P_SIZE"], as_index=False)["PS_SUPPKEY"] \
+                .nunique()
     total.columns = ["P_BRAND", "P_TYPE", "P_SIZE", "SUPPLIER_CNT"]
-    total = total.compute().sort_values(
+    total = total.sort_values(
         by=["SUPPLIER_CNT", "P_BRAND", "P_TYPE", "P_SIZE"],
         ascending=[False, True, True, True],
     )
@@ -857,12 +849,15 @@ def q16(root: str, storage_options: Dict):
     return total
 
 
-def q17(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    part = load_part(root, storage_options)
+def q17(root: str):
+    lineitem = load_lineitem(root)
+    part = load_part(root)
+
+    brand = "Brand#23"
+    container = "MED BOX"
 
     left = lineitem.loc[:, ["L_PARTKEY", "L_QUANTITY", "L_EXTENDEDPRICE"]]
-    right = part[((part["P_BRAND"] == "Brand#23") & (part["P_CONTAINER"] == "MED BOX"))]
+    right = part[((part["P_BRAND"] == brand) & (part["P_CONTAINER"] == container))]
     right = right.loc[:, ["P_PARTKEY"]]
     line_part_merge = left.merge(
         right, left_on="L_PARTKEY", right_on="P_PARTKEY", how="inner"
@@ -871,107 +866,96 @@ def q17(root: str, storage_options: Dict):
         :, ["L_QUANTITY", "L_EXTENDEDPRICE", "P_PARTKEY"]
     ]
     lineitem_filtered = lineitem.loc[:, ["L_PARTKEY", "L_QUANTITY"]]
-    lineitem_avg = (
-        lineitem_filtered.groupby(["L_PARTKEY"])
-        .L_QUANTITY.agg("mean")
-        .reset_index()
-        .rename(columns={"L_QUANTITY": "avg"})
-    )
+    lineitem_avg = lineitem_filtered.groupby(["L_PARTKEY"], as_index=False).mean()
+    lineitem_avg.columns = ["L_PARTKEY", "avg"]
     lineitem_avg["avg"] = 0.2 * lineitem_avg["avg"]
     lineitem_avg = lineitem_avg.loc[:, ["L_PARTKEY", "avg"]]
     total = line_part_merge.merge(
         lineitem_avg, left_on="P_PARTKEY", right_on="L_PARTKEY", how="inner"
     )
     total = total[total["L_QUANTITY"] < total["avg"]]
-    total = pandas.DataFrame(
-        {"avg_yearly": [(total["L_EXTENDEDPRICE"].sum() / 7.0).compute()]}
-    )
+    total = ps.DataFrame({"AVG_YEARLY": [total["L_EXTENDEDPRICE"].sum() / 7.0]})
 
     return total
 
 
-def q18(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    customer = load_customer(root, storage_options)
+def q18(root: str):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    customer = load_customer(root)
 
-    gb1 = lineitem.groupby("L_ORDERKEY")["L_QUANTITY"].sum().reset_index()
-    fgb1 = gb1[gb1.L_QUANTITY > 300]
+    quantity = 300
+    gb1 = lineitem.groupby("L_ORDERKEY", as_index=False)["L_QUANTITY"].sum()
+    fgb1 = gb1[gb1.L_QUANTITY > quantity]
     jn1 = fgb1.merge(orders, left_on="L_ORDERKEY", right_on="O_ORDERKEY")
     jn2 = jn1.merge(customer, left_on="O_CUSTKEY", right_on="C_CUSTKEY")
     gb2 = jn2.groupby(
         ["C_NAME", "C_CUSTKEY", "O_ORDERKEY", "O_ORDERDATE", "O_TOTALPRICE"],
+        as_index=False,
     )["L_QUANTITY"].sum()
-    total = (
-        gb2.compute()
-        .reset_index()
-        .sort_values(["O_TOTALPRICE", "O_ORDERDATE"], ascending=[False, True])
-    )
+    total = gb2.sort_values(["O_TOTALPRICE", "O_ORDERDATE"], ascending=[False, True])
     total = total.head(100)
 
     return total
 
 
-def q19(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    part = load_part(root, storage_options)
+def q19(root: str):
+    lineitem = load_lineitem(root)
+    part = load_part(root)
 
-    t1 = time.time()
-    Brand31 = "Brand#31"
-    Brand43 = "Brand#43"
-    SMBOX = "SM BOX"
-    SMCASE = "SM CASE"
-    SMPACK = "SM PACK"
-    SMPKG = "SM PKG"
-    MEDBAG = "MED BAG"
-    MEDBOX = "MED BOX"
-    MEDPACK = "MED PACK"
-    MEDPKG = "MED PKG"
-    LGBOX = "LG BOX"
-    LGCASE = "LG CASE"
-    LGPACK = "LG PACK"
-    LGPKG = "LG PKG"
-    DELIVERINPERSON = "DELIVER IN PERSON"
-    AIR = "AIR"
-    AIRREG = "AIRREG"
+    quantity1 = 4
+    quantity2 = 15
+    quantity3 = 26
+    brand1 = "Brand#31"
+    brand2 = "Brand#43"
+
     lsel = (
         (
-            ((lineitem.L_QUANTITY <= 36) & (lineitem.L_QUANTITY >= 26))
-            | ((lineitem.L_QUANTITY <= 25) & (lineitem.L_QUANTITY >= 15))
-            | ((lineitem.L_QUANTITY <= 14) & (lineitem.L_QUANTITY >= 4))
+            (
+                (lineitem.L_QUANTITY <= quantity3 + 10)
+                & (lineitem.L_QUANTITY >= quantity3)
+            )
+            | (
+                (lineitem.L_QUANTITY <= quantity2 + 10)
+                & (lineitem.L_QUANTITY >= quantity2)
+            )
+            | (
+                (lineitem.L_QUANTITY <= quantity1 + 10)
+                & (lineitem.L_QUANTITY >= quantity1)
+            )
         )
-        & (lineitem.L_SHIPINSTRUCT == DELIVERINPERSON)
-        & ((lineitem.L_SHIPMODE == AIR) | (lineitem.L_SHIPMODE == AIRREG))
+        & (lineitem.L_SHIPINSTRUCT == "DELIVER IN PERSON")
+        & ((lineitem.L_SHIPMODE == "AIR") | (lineitem.L_SHIPMODE == "AIRREG"))
     )
     psel = (part.P_SIZE >= 1) & (
         (
             (part.P_SIZE <= 5)
-            & (part.P_BRAND == Brand31)
+            & (part.P_BRAND == brand1)
             & (
-                (part.P_CONTAINER == SMBOX)
-                | (part.P_CONTAINER == SMCASE)
-                | (part.P_CONTAINER == SMPACK)
-                | (part.P_CONTAINER == SMPKG)
+                (part.P_CONTAINER == "SM BOX")
+                | (part.P_CONTAINER == "SM CASE")
+                | (part.P_CONTAINER == "SM PACK")
+                | (part.P_CONTAINER == "SM PKG")
             )
         )
         | (
             (part.P_SIZE <= 10)
-            & (part.P_BRAND == Brand43)
+            & (part.P_BRAND == brand2)
             & (
-                (part.P_CONTAINER == MEDBAG)
-                | (part.P_CONTAINER == MEDBOX)
-                | (part.P_CONTAINER == MEDPACK)
-                | (part.P_CONTAINER == MEDPKG)
+                (part.P_CONTAINER == "MED BAG")
+                | (part.P_CONTAINER == "MED BOX")
+                | (part.P_CONTAINER == "MED PACK")
+                | (part.P_CONTAINER == "MED PKG")
             )
         )
         | (
             (part.P_SIZE <= 15)
-            & (part.P_BRAND == Brand43)
+            & (part.P_BRAND == brand2)
             & (
-                (part.P_CONTAINER == LGBOX)
-                | (part.P_CONTAINER == LGCASE)
-                | (part.P_CONTAINER == LGPACK)
-                | (part.P_CONTAINER == LGPKG)
+                (part.P_CONTAINER == "LG BOX")
+                | (part.P_CONTAINER == "LG CASE")
+                | (part.P_CONTAINER == "LG PACK")
+                | (part.P_CONTAINER == "LG PKG")
             )
         )
     )
@@ -979,59 +963,55 @@ def q19(root: str, storage_options: Dict):
     fpart = part[psel]
     jn = flineitem.merge(fpart, left_on="L_PARTKEY", right_on="P_PARTKEY")
     jnsel = (
-        (
-            (jn.P_BRAND == Brand31)
-            & (
-                (jn.P_CONTAINER == SMBOX)
-                | (jn.P_CONTAINER == SMCASE)
-                | (jn.P_CONTAINER == SMPACK)
-                | (jn.P_CONTAINER == SMPKG)
-            )
-            & (jn.L_QUANTITY >= 4)
-            & (jn.L_QUANTITY <= 14)
-            & (jn.P_SIZE <= 5)
+        (jn.P_BRAND == brand1)
+        & (
+            (jn.P_CONTAINER == "SM BOX")
+            | (jn.P_CONTAINER == "SM CASE")
+            | (jn.P_CONTAINER == "SM PACK")
+            | (jn.P_CONTAINER == "SM PKG")
         )
-        | (
-            (jn.P_BRAND == Brand43)
-            & (
-                (jn.P_CONTAINER == MEDBAG)
-                | (jn.P_CONTAINER == MEDBOX)
-                | (jn.P_CONTAINER == MEDPACK)
-                | (jn.P_CONTAINER == MEDPKG)
-            )
-            & (jn.L_QUANTITY >= 15)
-            & (jn.L_QUANTITY <= 25)
-            & (jn.P_SIZE <= 10)
+        & (jn.L_QUANTITY >= quantity1)
+        & (jn.L_QUANTITY <= quantity1 + 10)
+        & (jn.P_SIZE <= 5)
+        | (jn.P_BRAND == brand2)
+        & (
+            (jn.P_CONTAINER == "MED BAG")
+            | (jn.P_CONTAINER == "MED BOX")
+            | (jn.P_CONTAINER == "MED PACK")
+            | (jn.P_CONTAINER == "MED PKG")
         )
-        | (
-            (jn.P_BRAND == Brand43)
-            & (
-                (jn.P_CONTAINER == LGBOX)
-                | (jn.P_CONTAINER == LGCASE)
-                | (jn.P_CONTAINER == LGPACK)
-                | (jn.P_CONTAINER == LGPKG)
-            )
-            & (jn.L_QUANTITY >= 26)
-            & (jn.L_QUANTITY <= 36)
-            & (jn.P_SIZE <= 15)
+        & (jn.L_QUANTITY >= quantity2)
+        & (jn.L_QUANTITY <= quantity2 + 10)
+        & (jn.P_SIZE <= 10)
+        | (jn.P_BRAND == brand2)
+        & (
+            (jn.P_CONTAINER == "LG BOX")
+            | (jn.P_CONTAINER == "LG CASE")
+            | (jn.P_CONTAINER == "LG PACK")
+            | (jn.P_CONTAINER == "LG PKG")
         )
+        & (jn.L_QUANTITY >= quantity3)
+        & (jn.L_QUANTITY <= quantity3 + 10)
+        & (jn.P_SIZE <= 15)
     )
     jn = jn[jnsel]
-    total = (jn.L_EXTENDEDPRICE * (1.0 - jn.L_DISCOUNT)).sum()
+    result_value = (jn.L_EXTENDEDPRICE * (1.0 - jn.L_DISCOUNT)).sum()
+    result_df = ps.DataFrame({"REVENUE": [result_value]})
 
-    return total
+    return result_df
 
 
-def q20(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    part = load_part(root, storage_options)
-    nation = load_nation(root, storage_options)
-    partsupp = load_partsupp(root, storage_options)
-    supplier = load_supplier(root, storage_options)
+def q20(root: str):
+    lineitem = load_lineitem(root)
+    part = load_part(root)
+    nation = load_nation(root)
+    partsupp = load_partsupp(root)
+    supplier = load_supplier(root)
 
-    date1 = datetime.strptime("1996-01-01", "%Y-%m-%d")
-    date2 = datetime.strptime("1997-01-01", "%Y-%m-%d")
-    psel = part.P_NAME.str.startswith("azure")
+    p_name = "azure"
+    date1 = pd.Timestamp("1996-01-01")
+    date2 = date1 + pd.DateOffset(years=1)
+    psel = part.P_NAME.str.startswith(p_name)
     nsel = nation.N_NAME == "JORDAN"
     lsel = (lineitem.L_SHIPDATE >= date1) & (lineitem.L_SHIPDATE < date2)
     fpart = part[psel]
@@ -1043,28 +1023,26 @@ def q20(root: str, storage_options: Dict):
         left_on=["PS_PARTKEY", "PS_SUPPKEY"],
         right_on=["L_PARTKEY", "L_SUPPKEY"],
     )
-    gb = (
-        jn2.groupby(["PS_PARTKEY", "PS_SUPPKEY", "PS_AVAILQTY"])["L_QUANTITY"]
-        .sum()
-        .reset_index()
-    )
+    gb = jn2.groupby(["PS_PARTKEY", "PS_SUPPKEY", "PS_AVAILQTY"], as_index=False)[
+        "L_QUANTITY"
+    ].sum()
     gbsel = gb.PS_AVAILQTY > (0.5 * gb.L_QUANTITY)
     fgb = gb[gbsel]
     jn3 = fgb.merge(supplier, left_on="PS_SUPPKEY", right_on="S_SUPPKEY")
     jn4 = fnation.merge(jn3, left_on="N_NATIONKEY", right_on="S_NATIONKEY")
     jn4 = jn4.loc[:, ["S_NAME", "S_ADDRESS"]]
-    total = jn4.compute().sort_values("S_NAME").drop_duplicates()
+    total = jn4.sort_values("S_NAME").drop_duplicates()
 
     return total
 
 
-def q21(root: str, storage_options: Dict):
-    lineitem = load_lineitem(root, storage_options)
-    orders = load_orders(root, storage_options)
-    supplier = load_supplier(root, storage_options)
-    nation = load_nation(root, storage_options)
+def q21(root: str):
+    lineitem = load_lineitem(root)
+    orders = load_orders(root)
+    supplier = load_supplier(root)
+    nation = load_nation(root)
 
-    t1 = time.time()
+    nation_name = "SAUDI ARABIA"
     lineitem_filtered = lineitem.loc[
         :, ["L_ORDERKEY", "L_SUPPKEY", "L_RECEIPTDATE", "L_COMMITDATE"]
     ]
@@ -1072,9 +1050,8 @@ def q21(root: str, storage_options: Dict):
     # Exists
     lineitem_orderkeys = (
         lineitem_filtered.loc[:, ["L_ORDERKEY", "L_SUPPKEY"]]
-        .groupby("L_ORDERKEY")["L_SUPPKEY"]
+        .groupby("L_ORDERKEY", as_index=False)["L_SUPPKEY"]
         .nunique()
-        .reset_index()
     )
     lineitem_orderkeys.columns = ["L_ORDERKEY", "nunique_col"]
     lineitem_orderkeys = lineitem_orderkeys[lineitem_orderkeys["nunique_col"] > 1]
@@ -1093,7 +1070,8 @@ def q21(root: str, storage_options: Dict):
 
     # Not Exists: Check the exists condition isn't still satisfied on the output.
     lineitem_orderkeys = (
-        lineitem_filtered.groupby("L_ORDERKEY")["L_SUPPKEY"].nunique().reset_index()
+        lineitem_filtered.groupby("L_ORDERKEY", as_index=False)["L_SUPPKEY"]
+        .nunique()
     )
     lineitem_orderkeys.columns = ["L_ORDERKEY", "nunique_col"]
     lineitem_orderkeys = lineitem_orderkeys[lineitem_orderkeys["nunique_col"] == 1]
@@ -1118,43 +1096,39 @@ def q21(root: str, storage_options: Dict):
     )
     total = total.loc[:, ["S_NATIONKEY", "S_NAME"]]
     nation_filtered = nation.loc[:, ["N_NAME", "N_NATIONKEY"]]
-    nation_filtered = nation_filtered[nation_filtered["N_NAME"] == "SAUDI ARABIA"]
+    nation_filtered = nation_filtered[nation_filtered["N_NAME"] == nation_name]
     total = total.merge(
         nation_filtered, left_on="S_NATIONKEY", right_on="N_NATIONKEY", how="inner"
     )
     total = total.loc[:, ["S_NAME"]]
-    total = total.groupby("S_NAME").size().reset_index()
+    total = total.groupby("S_NAME", as_index=False).size().reset_index()
     total.columns = ["S_NAME", "NUMWAIT"]
-    total = total.compute().sort_values(
-        by=[
-            "NUMWAIT",
-            "S_NAME",
-        ],
-        ascending=[
-            False,
-            True,
-        ],
-    )
+    total = total.sort_values(by=["NUMWAIT", "S_NAME"], ascending=[False, True])
+    total = total.head(100)
 
     return total
 
 
-def q22(root: str, storage_options: Dict):
-    customer = load_customer(root, storage_options)
-    orders = load_orders(root, storage_options)
+def q22(root: str):
+    customer = load_customer(root)
+    orders = load_orders(root)
 
+    I1 = "13"
+    I2 = "31"
+    I3 = "23"
+    I4 = "29"
+    I5 = "30"
+    I6 = "18"
+    I7 = "17"
     customer_filtered = customer.loc[:, ["C_ACCTBAL", "C_CUSTKEY"]]
     customer_filtered["CNTRYCODE"] = customer["C_PHONE"].str.slice(0, 2)
     customer_filtered = customer_filtered[
         (customer["C_ACCTBAL"] > 0.00)
-        & customer_filtered["CNTRYCODE"].isin(
-            ["13", "31", "23", "29", "30", "18", "17"]
-        )
+        & customer_filtered["CNTRYCODE"].isin([I1, I2, I3, I4, I5, I6, I7])
     ]
     avg_value = customer_filtered["C_ACCTBAL"].mean()
     customer_filtered = customer_filtered[customer_filtered["C_ACCTBAL"] > avg_value]
-    # Select only the keys that don't match by performing a left join and only
-    # selecting columns with an N/A value
+    # Select only the keys that don't match by performing a left join and only selecting columns with an na value
     orders_filtered = orders.loc[:, ["O_CUSTKEY"]].drop_duplicates()
     customer_keys = customer_filtered.loc[:, ["C_CUSTKEY"]].drop_duplicates()
     customer_selected = customer_keys.merge(
@@ -1166,20 +1140,13 @@ def q22(root: str, storage_options: Dict):
         customer_filtered, on="C_CUSTKEY", how="inner"
     )
     customer_selected = customer_selected.loc[:, ["CNTRYCODE", "C_ACCTBAL"]]
-    agg1 = customer_selected.groupby(["CNTRYCODE"]).size().reset_index()
+    agg1 = customer_selected.groupby(["CNTRYCODE"], as_index=False).size().reset_index()
     agg1.columns = ["CNTRYCODE", "NUMCUST"]
-    agg2 = customer_selected.groupby(["CNTRYCODE"]).C_ACCTBAL.agg("sum").reset_index()
-    agg2 = agg2.rename(columns={"C_ACCTBAL": "TOTACCTBAL"})
+    agg2 = customer_selected.groupby(["CNTRYCODE"], as_index=False).sum()
+    agg2.columns = ["CNTRYCODE", "TOTACCTBAL"]
     total = agg1.merge(agg2, on="CNTRYCODE", how="inner")
-    total = total.compute().sort_values(
-        by=[
-            "CNTRYCODE",
-        ],
-        ascending=[
-            True,
-        ],
-    )
-    
+    total = total.sort_values(by=["CNTRYCODE"], ascending=[True])
+
     return total
 
 
@@ -1256,144 +1223,58 @@ query_to_runner = {
 }
 
 
-def get_query_answer(query: int, base_dir: str = ANSWERS_BASE_DIR) -> PandasDF:
-    answer_df = pandas.read_csv(
-        os.path.join(base_dir, f"q{query}.out"),
-        sep="|",
-        parse_dates=True,
-        infer_datetime_format=True,
-    )
-    return answer_df.rename(columns=lambda x: x.strip())
-
-
-def test_results(q_num: int, result_df: PandasDF):
-    answer = get_query_answer(q_num)
-
-    for column_index in range(len(answer.columns)):
-        column_name = answer.columns[column_index]
-        column_data_type = answer.iloc[:, column_index].dtype
-
-        s1 = result_df.iloc[:, column_index]
-        s2 = answer.iloc[:, column_index]
-
-        if column_data_type.name == "object":
-            s1 = s1.astype("string").apply(lambda x: x.strip())
-            s2 = s2.astype("string").apply(lambda x: x.strip())
-
-        pandas.testing.assert_series_equal(
-            left=s1,
-            right=s2,
-            check_index=False,
-            check_names=False,
-            check_exact=False,
-            rtol=1e-2,
-        )
-
-
-def run_queries(path, 
-        storage_options, 
-        client, 
-        queries,
-        log_time=True,
-        include_io=False,
-        test_result=False,
-        print_result=False
-    ):
-    total_start = time.time()
-    print("Start data loading")
+def run_queries(
+    path,
+    queries,
+    log_time=True,
+    print_result=False,
+    include_io=False,
+):
+    version = pyspark.__version__
+    data_start_time = time.time()
     for query in queries:
         loaders = query_to_loaders[query]
         for loader in loaders:
-            loader(path, storage_options=storage_options=False)
-    print(f"Data loading time (s): {time.time() - total_start}")
-
-    total_start = time.time()
-    print("Start data persisting")
-    for table_name in dataset_dict:
-        df = dataset_dict[table_name]
-        start = time.time()
-        df = client.persist(df)
-        wait(df)
-        dataset_dict[table_name] = df
-        print(f"{table_name} persisting time (s): {time.time() - start}")
-    print(f"Total data persisting time (s): {time.time() - total_start}")
+            loader(path)
+    print(f"Total data loading time (s): {time.time() - data_start_time}")
 
     total_start = time.time()
     for query in queries:
         try:
-            t1 = time.time()
-            result = query_to_runner[query](path, storage_options)
-            dur = time.time() - t1
+            start_time = time.time()
+            result = query_to_runner[query](path)
+            result = result.to_pandas()
+            without_io_time = time.time() - start_time
             success = True
-            if test_result:
-                test_results(query, result)
             if print_result:
-                print(result)
+                print_result_fn("pyspark_pandas", result, query)
         except Exception as e:
             print("".join(traceback.TracebackException.from_exception(e).format()))
-            dur = 0.0
+            without_io_time = 0.0
             success = False
         finally:
-            if log_time:
-                append_row("dask", query, dur, dask.__version__, success)
+            pass
+        if log_time:
+            log_time_fn("pyspark_pandas", query, version=version, without_io_time=without_io_time, success=success)
     print(f"Total query execution time (s): {time.time() - total_start}")
 
 
 def main():
+    global spark
     parser = argparse.ArgumentParser(description="TPC-H benchmark.")
-    parser.add_argument(
-        "--path", type=str, required=True, help="Path to the TPC-H dataset."
-    )
-    parser.add_argument(
-        "--storage_options",
-        type=str,
-        required=False,
-        help="Path to the storage options json file.",
-    )
-    parser.add_argument(
-        "--queries",
-        type=int,
-        nargs="+",
-        required=False,
-        help="Comma separated TPC-H queries to run.",
-    )
-    parser.add_argument(
-        "--endpoint",
-        type=str,
-        required=False,
-        help="The endpoint of existing Dask cluster."
-    )
-    parser.add_argument(
-        "--log_time",
-        action="store_true",
-        help="log time metrics or not.",
-    )
-    parser.add_argument(
-        "--include_io",
-        action="store_true",
-        help="include IO or not.",
-    )
-    parser.add_argument(
-        "--test_result",
-        action="store_true",
-        help="test results with official answers.",
-    )
-    parser.add_argument(
-        "--print_result",
-        action="store_true",
-        help="print result.",
-    )
+    parser.add_argument("--master", type=str, help="Spark master URI")
+
+    # aws settings
+    parser.add_argument("--account", type=str, help="AWS access id")
+    parser.add_argument("--key", type=str, help="AWS secret access key")
+    parser.add_argument("--endpoint", type=str, help="AWS region endpoint related to your S3")
+
+    parser.add_argument("--executor_cores", type=int, default=32, help='Number of cores for each Spark executor')
+    parser.add_argument("--executor_memory", type=str, default="64G", help='Memory size for each Spark executor')
+    parser = parse_common_arguments(parser)
     args = parser.parse_args()
-
-    # path to TPC-H data in parquet.
+    path: str = args.path
     print(f"Path: {args.path}")
-
-    # credentials to access the datasource.
-    storage_options = {}
-    if args.storage_options is not None:
-        with open(args.storage_options, "r") as fp:
-            storage_options = json.load(fp)
-    print(f"Storage options: {storage_options}")
 
     queries = list(range(1, 23))
     if args.queries is not None:
@@ -1401,20 +1282,35 @@ def main():
     print(f"Queries to run: {queries}")
     print(f"Include IO: {args.include_io}")
 
-    if args.endpoint == "local" or args.endpoint is None:
-        from dask.distributed import LocalCluster
-        client = LocalCluster()
-    elif args.endpoint:
-        client = Client(args.endpoint)
+    account = args.account
+    key = args.key
+
+    spark = SparkSession\
+        .builder\
+        .appName("PySpark tpch query")\
+        .master(args.master) \
+        .config("spark.executor.cores", args.executor_cores) \
+        .config("spark.executor.memory", args.executor_memory) \
+        .config('spark.ui.showConsoleProgress', 'false') \
+        .getOrCreate()
     
-    run_queries(args.path, 
-        storage_options, 
-        client, queries, 
-        args.log_time, 
-        args.include_io, 
-        args.test_result, 
-        args.print_result
-    )
+    
+    spark.sparkContext.setLogLevel("ERROR")
+
+    if "s3://" in path:
+        conf = spark.sparkContext._jsc.hadoopConfiguration()
+        conf.set("fs.s3a.access.key", account)
+        conf.set("fs.s3a.secret.key", key)
+        conf.set("fs.s3a.endpoint", args.endpoint)
+        path = path.replace("s3://", "s3a://")
+        # if you are using Hadoop or S3 on a cloud, you should add the following config
+        spark.conf.set("spark.jars.packages",'org.apache.spark:spark-hadoop-cloud_2.12:3.5.0')
+    
+    run_queries(path, 
+                queries, 
+                args.log_time, 
+                args.print_result,
+                args.include_io)
 
 
 if __name__ == "__main__":
